+++ pkg/tsengine/commands/log.go
