+++ pkg/tsengine/commands/strings.go
