+++ pkg/tsengine/compiler/definition.go
