+++ pkg/tsengine/runtime_manager.go
@@ -9,12 +9,12 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/direktiv/direktiv/pkg/commands"
 	"github.com/direktiv/direktiv/pkg/core"
 	"github.com/direktiv/direktiv/pkg/engine"
 	"github.com/direktiv/direktiv/pkg/runtime"
-	"github.com/direktiv/direktiv/pkg/state"
 	"github.com/direktiv/direktiv/pkg/tracing"
+	"github.com/direktiv/direktiv/pkg/tsengine/commands"
+	"github.com/direktiv/direktiv/pkg/tsengine/state"
 	"github.com/dop251/goja"
 	"github.com/google/uuid"
 )
