+++ pkg/tsengine/server_test.go
@@ -13,13 +13,13 @@ import (
 	"testing"
 
 	"github.com/caarlos0/env/v11"
-	"github.com/direktiv/direktiv/pkg/compiler"
 	"github.com/direktiv/direktiv/pkg/database"
 	"github.com/direktiv/direktiv/pkg/datastore"
 	"github.com/direktiv/direktiv/pkg/datastore/datastoresql"
 	"github.com/direktiv/direktiv/pkg/filestore"
 	"github.com/direktiv/direktiv/pkg/filestore/filestoresql"
 	"github.com/direktiv/direktiv/pkg/tsengine"
+	"github.com/direktiv/direktiv/pkg/tsengine/compiler"
 	"github.com/direktiv/direktiv/pkg/utils"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
